Calculator

+++ src/home_work_3/calcs/api/ICalculator.java
@@ -2,7 +2,7 @@ package home_work_3.calcs.api;
 
 public interface ICalculator {
 
-    public double plus(double a, int b, double c);
+    public double plus(double a, double b, double c);
 
     public double pow(double c, double d);
